Create `Backbone` base class
